# backend/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/api/math/validate/question/positive": {
      "post": {
        "summary": "Validate Positive",
        "operationId": "validate_positive_api_math_validate_question_positive_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_validate_positive_api_math_validate_question_positive_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/math/validate/question/length": {
      "post": {
        "summary": "Validate Question Length",
        "operationId": "validate_question_length_api_math_validate_question_length_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_validate_question_length_api_math_validate_question_length_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/math/validate/test/length": {
      "post": {
        "summary": "Validate Test Length",
        "operationId": "validate_test_length_api_math_validate_test_length_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_validate_test_length_api_math_validate_test_length_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/math/validate/test/grade": {
      "post": {
        "summary": "Validate Grade",
        "operationId": "validate_grade_api_math_validate_test_grade_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_validate_grade_api_math_validate_test_grade_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/math/check/teacher": {
      "post": {
        "summary": "Teacher Check",
        "operationId": "teacher_check_api_math_check_teacher_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_teacher_check_api_math_check_teacher_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_teacher_check_api_math_check_teacher_post": {
        "title": "Body_teacher_check_api_math_check_teacher_post",
        "required": [
          "test",
          "teacher_name",
          "teacher_id"
        ],
        "type": "object",
        "properties": {
          "test": {
            "title": "Test",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Question"
            }
          },
          "teacher_name": {
            "title": "Teacher Name",
            "type": "string"
          },
          "teacher_id": {
            "title": "Teacher Id",
            "type": "string"
          }
        }
      },
      "Body_validate_grade_api_math_validate_test_grade_post": {
        "title": "Body_validate_grade_api_math_validate_test_grade_post",
        "required": [
          "test"
        ],
        "type": "object",
        "properties": {
          "test": {
            "title": "Test",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Question"
            }
          },
          "max_grade": {
            "title": "Max Grade",
            "type": "integer",
            "default": 100
          }
        }
      },
      "Body_validate_positive_api_math_validate_question_positive_post": {
        "title": "Body_validate_positive_api_math_validate_question_positive_post",
        "required": [
          "q",
          "bruh"
        ],
        "type": "object",
        "properties": {
          "q": {
            "$ref": "#/components/schemas/Question"
          },
          "bruh": {
            "title": "Bruh",
            "type": "integer"
          }
        }
      },
      "Body_validate_question_length_api_math_validate_question_length_post": {
        "title": "Body_validate_question_length_api_math_validate_question_length_post",
        "required": [
          "q"
        ],
        "type": "object",
        "properties": {
          "q": {
            "$ref": "#/components/schemas/Question"
          },
          "max_length": {
            "title": "Max Length",
            "type": "integer",
            "default": 10
          }
        }
      },
      "Body_validate_test_length_api_math_validate_test_length_post": {
        "title": "Body_validate_test_length_api_math_validate_test_length_post",
        "required": [
          "test"
        ],
        "type": "object",
        "properties": {
          "test": {
            "title": "Test",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Question"
            }
          }
        }
      },
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "Question": {
        "title": "Question",
        "required": [
          "q",
          "value"
        ],
        "type": "object",
        "properties": {
          "q": {
            "title": "Q",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "integer"
          }
        }
      },
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    }
  }
}